Do not use Intent.resolveActivity/resolveActivityInfo to verify if a URL is valid. (#101)
Google changed the behaviour of Intent.resolveActivity/resolveActivityInfo when targeting recent API levels in order to prevent apps abusing it to scan the device for apps. As customers begin to target later API levels, this is now turning up in the field.

They now only allow this method to work now with apps you have explicitly requested previously in your manifest. This can't apply to our usage of it, which is a sanity check to prevent a helpful message in the event the app developer has integrated our SDK incorrectly. This patch fixes this to handle it by catching the exception instead.

Workaround in the meantime before that fix is released, communicated to client, is to use bit.ly to bounce the URL to YouTube, since web URLs (not handled by any other app on device other than browser) do not cause Intent.resolveActivity/resolveActivityInfo to return null.

See details about Google's change at:

https://stackoverflow.com/questions/62535856/intent-resolveactivity-returns-null-in-api-30

https://medium.com/androiddevelopers/package-visibility-in-android-11-cc857f221cd9