Create Github Actions for CI and Deploy (#102)
This sets up GitHub Actions-powered CI and Deployment workflows, pretty similarly to how they've been done on judo-compose.

CI runs automatically on PRs and on master.

Deploy must be triggered from the Actions tab when master (or a release branch) is ready to deploy.

Deploy builds the AARs, and then commits and pushes them to the judo-maven repo. This means that going forward customers will have to use the [judo-maven](https://github.com/judoapp/judo-maven) Maven repository instead of the one hosted directly in this repo, which will no longer be updated.

Bumped Gradle version, although AGP has remained at the same version for now.

This change also officially ends the use of gitflow on this repo. The `develop` branch has been removed.